requirements.txt
alembic==1.1.0
certifi==2019.6.16
cffi==1.12.3
Click==7.0
django-environ==0.4.5
Flask==1.1.1
Flask-Login==0.4.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.1.0
MarkupSafe==1.1.1
marshmallow==3.0.3
psycopg2==2.8.3
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.8.0
python-editor==1.0.4
six==1.12.0
SQLAlchemy==1.3.8
Werkzeug==0.15.6
